requirements.txt
face_recognition
pillow